optmize shm. add raspi async and multiprocess queue socket

+++ webserver/raspi/socket_mp-h264.py
@@ -1,5 +1,9 @@
 import platform
 import os
+import  queue
+import socket
+import time
+
 system = platform.system()
 
 def install_loop():
@@ -19,7 +23,8 @@ install_loop()
 
 # Add the directory containing FFmpeg DLLs
 ffmpeg_bin = r"C:\ffmpeg\bin"
-os.add_dll_directory(ffmpeg_bin)
+if system == 'Windows' and os.path.exists(ffmpeg_bin):
+    os.add_dll_directory(ffmpeg_bin)
 
 import struct
 import asyncio
@@ -27,7 +32,7 @@ import cv2
 import contextlib
 from zlib import crc32
 import multiprocessing
-
+import signal
 ''' 
     UDP Configuration
 '''
@@ -36,12 +41,34 @@ EC2_UDP_PORT = 8086
 MAX_UDP_PACKET_SIZE = 60000  # Max safe UDP payload size
 CAMERA_INDEX = 0             # Default camera index
 
+''' Global Variable '''
+frame_id_counter = 0
+
+# Shutdown event
+keep_running = True
+def handle_exit_signal(signum, frame):
+    print(f"\nReceived signal {signum}, shutting down...")
+    global keep_running
+    keep_running = False
+
+signal.signal(signal.SIGINT, handle_exit_signal)
+signal.signal(signal.SIGTERM, handle_exit_signal)
+
 class UDPSender(asyncio.DatagramProtocol):
     def __init__(self):
         self.transport = None
 
     def connection_made(self, transport):
         self.transport = transport
+        sock: socket.socket = transport.get_extra_info('socket')
+
+        default_sndbuf = sock.getsockopt(socket.SOL_SOCKET, socket.SO_SNDBUF)
+        print(f"Default SO_sndbuf: {default_sndbuf} bytes")
+
+        sock.setsockopt(socket.SOL_SOCKET, socket.SO_SNDBUF, 32 * 1024 * 1024)
+
+        new_sndbuf = sock.getsockopt(socket.SOL_SOCKET, socket.SO_SNDBUF)
+        print(f"new SO_sndbuf: {new_sndbuf} bytes")
 
     def datagram_received(self, data, addr):
         print(f"Received from {addr}: ", data.decode())
@@ -56,8 +83,6 @@ class UDPSender(asyncio.DatagramProtocol):
     def connection_lost(self, exc):
         print("Connection closed")
 
-''' Global Variable '''
-frame_id_counter = 0
 
 class VideoStream:
     def __init__(self, frame_queue:multiprocessing.Queue, src=0, desired_width=680,):
@@ -73,10 +98,9 @@ class VideoStream:
         self.ret, self.frame = self.cap.read()
         self.frame_queue = frame_queue
         self.desired_width = desired_width
-        self.stopped = False
+        self.running = True
     
     async def start(self):
-        self.running = True
         loop = asyncio.get_running_loop()
 
         while self.running:
@@ -87,23 +111,26 @@ class VideoStream:
                     continue
 
                 #frame = cv2.resize(frame, (self.desired_width, int(frame.shape[0] * self.desired_width / frame.shape[1])))
-                if not  self.frame_queue.full():
+                if not self.frame_queue.full():
                     self.frame_queue.put_nowait(frame.copy())
                 else:
-                    print("full")
+                    print("frame_queue full")
+                    if not keep_running:
+                        break
 
                 await asyncio.sleep(0)
+            except asyncio.CancelledError:
+                break
             except Exception as e:
-                print("videostream start error: {e}")
-            
+                print(f"VideoStream start error: {e}")
+        
+        print("exited...")
+
     def stop(self):
-        self.stopped = True
+        self.running = False
         self.cap.release()
 
 async def encode_video(frame_queue: multiprocessing.Queue, encode_queue: multiprocessing.Queue):
-    install_loop()
-    loop = asyncio.get_running_loop()
-
     import av
     encoder = av.CodecContext.create('libx264', 'w')
     encoder.width = 1280
@@ -113,13 +140,9 @@ async def encode_video(frame_queue: multiprocessing.Queue, encode_queue: multipr
     encoder.framerate = 30 
     encoder.options = {'tune': 'zerolatency'} 
 
-    try:
-        while True:
-            if not frame_queue.empty():
-                frame = frame_queue.get()
-            else:
-                await asyncio.sleep(0.02)
-                continue
+    while True:
+        try:
+            frame = frame_queue.get()
 
             img_yuv = cv2.cvtColor(frame, cv2.COLOR_BGR2YUV_I420)
             video_frame = av.VideoFrame.from_ndarray(img_yuv, format='yuv420p')
@@ -132,10 +155,13 @@ async def encode_video(frame_queue: multiprocessing.Queue, encode_queue: multipr
             if not encode_queue.full():
                 encode_queue.put_nowait(encoded_packet_bytes)
             else:
-                print("full")
-    except Exception as e:
-        print(f"error: {e}")
-
+                print("encode_queue full")
+        except InterruptedError:
+            break
+        except asyncio.CancelledError:
+            break
+        except Exception as e:
+            print(f"error: {e}")
         
 async def send_frame(protocol: UDPSender, encoded_frame: bytes, addr: tuple[str, int]):
     global frame_id_counter
@@ -158,11 +184,21 @@ async def send_frame(protocol: UDPSender, encoded_frame: bytes, addr: tuple[str,
         protocol.send(header + chunk, addr)
         
 def async_encode(frame_queue: multiprocessing.Queue, encode_queue: multiprocessing.Queue):
-    asyncio.run(encode_video(frame_queue,encode_queue))
+    install_loop()
+    try:
+        asyncio.run(encode_video(frame_queue, encode_queue))
+    except KeyboardInterrupt:
+        print("exiting...")
+    except SystemExit:
+        print("exiting...")
+
 
 async def main():
-    frame_queue  = multiprocessing.Queue(60)
-    encode_queue = multiprocessing.Queue(60)
+    if system == 'Windows':
+        multiprocessing.set_start_method('spawn')
+
+    frame_queue  = multiprocessing.Queue(120)
+    encode_queue = multiprocessing.Queue(120)
     vs = VideoStream(frame_queue)
     capture_task = asyncio.create_task(vs.start())
 
@@ -177,29 +213,70 @@ async def main():
         remote_addr=(EC2_UDP_IP, EC2_UDP_PORT)
     )
 
+    frame_count = 0
+    prev_time = time.monotonic()
+
     try:
-        while True:
-            if not encode_queue.empty():
-                encoded_frame = await loop.run_in_executor(None, encode_queue.get)
+        while keep_running:
+            try:
+                encoded_frame = await loop.run_in_executor(None, lambda: encode_queue.get(timeout=5))
                 await send_frame(protocol, encoded_frame, (EC2_UDP_IP, EC2_UDP_PORT))
-            else:
-                await asyncio.sleep(0.02)
+
+                frame_count += 1
+                now = time.monotonic()
+                if now - prev_time >= 1.0:
+                    fps = frame_count / (now - prev_time)
+                    print(f"FPS: {fps:.2f}")
+                    frame_count = 0
+                    prev_time = now
+
+                await asyncio.sleep(0)
+            except queue.Empty:
+                continue
+            except asyncio.CancelledError:
+                break
     except asyncio.CancelledError:
         pass
     except KeyboardInterrupt:
         pass
     finally:
+        print("cancelling task......")
+        capture_task.cancel()
         vs.stop()
-        capture_task.cancel()
-        encode_process.terminate()
-        encode_process.join()
-        capture_task.cancel()
+        transport.close()        
+        cv2.destroyAllWindows()
+
+        try:
+            if encode_process.is_alive():
+                os.kill(encode_process.pid, signal.SIGINT)
+        except Exception as e:
+            print(e)
+
+        encode_process.join(2)
+        if encode_process.is_alive():
+            print("terminating encode_proccess")
+            encode_process.terminate()
+        encode_process.join(3)
+        if encode_process.is_alive():
+            print("killing encode_proccess")
+            encode_process.kill()
+        encode_process.join(2)
+
         try:
             with contextlib.suppress(asyncio.CancelledError):
+                print("awaiting capture_task")
                 await capture_task
+                print("capture_task ended")
+                frame_queue.cancel_join_thread()
+                encode_queue.cancel_join_thread()
+                frame_queue.close()
+                encode_queue.close()
         except Exception as e:
             print(f"Error occurred while canceling stream task: {e}")
 
 if __name__ == '__main__':
-    asyncio.run(main())
+    try:
+        asyncio.run(main())
+    except KeyboardInterrupt as e:
+        print("Program interrupted by user. Exiting...")
